style: Mudança de lugar do middleware de autenticação JWT
