+++ src/server/routes/Route.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { LoginSchema } from '../middleware/LoginAuth';
+import { loginValidation } from '../middleware/LoginSchema';
 import { bodyValidation } from '../middleware/BodyValidation';
 import { VerifySignUp } from '../middleware/VerifySignUp';
 import { signUp } from '../controllers/SignUp'
@@ -7,7 +7,7 @@ import { login } from '../controllers/Login';
 
 export const router = Router();
 
-router.post('/login', bodyValidation(LoginSchema.loginValidation), login);
+router.post('/login', bodyValidation(loginValidation), login);
 
 router.post('/signup', [ bodyValidation(VerifySignUp.signUpValidation), 
     VerifySignUp.checkDuplicateEmail, VerifySignUp.checkDuplicateUsername ], signUp)
